petapp_backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.8.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
